pyproject.toml
[project]
name = "auto-file-extractor"
version = "0.1.0"
description = ""
authors = [
    {name = "lucas-almeida-1",email = "luquinhas.al@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rarfile (>=4.2,<5.0)",
    "py7zr (>=1.0.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
